+++ src/realm/query.cpp
@@ -1982,7 +1982,7 @@ Query Query::operator&&(const Query& q)
 Query Query::operator!()
 {
     if (!root_node())
-        throw util::runtime_error("negation of empty query is not supported");
+        throw ExceptionForStatus(ErrorCodes::InvalidQuery, "negation of empty query is not supported");
     Query q(m_table);
     q.Not();
     q.and_query(*this);