+++ test/test_shared.cpp
@@ -2340,7 +2340,7 @@ TEST(Shared_EncryptionKeyCheck_2)
     try {
         DBRef sg_2 = DB::create(path, false, DBOptions(crypt_key(true)));
     }
-    catch (std::runtime_error&) {
+    catch (std::exception&) {
         ok = true;
     }
     CHECK(ok);
@@ -3488,12 +3488,12 @@ TEST(Shared_ConstObjectIterator)
     t4->clear();
     auto i5(i4);
     // dereferencing an invalid iterator will throw
-    CHECK_THROW(*i5, std::logic_error);
+    CHECK_THROW(*i5, realm::ExceptionForStatus);
     // but moving it will not, it just stays invalid
     ++i5;
     i5 += 3;
     // so, should still throw
-    CHECK_THROW(*i5, std::logic_error);
+    CHECK_THROW(*i5, realm::ExceptionForStatus);
     CHECK(i5 == t4->end());
 }